move serialization tests
